Remove `bottle :unneeded` to silence deprecation warning.
This prevents Homebrew from printing a deprecation warning on every invocation of `brew`:

```
Warning: Calling bottle :unneeded is deprecated! There is no replacement.
Please report this issue to the alphagov/gds tap (not Homebrew/brew or Homebrew/core):
  /usr/local/Homebrew/Library/Taps/alphagov/homebrew-gds/Formula/govuk-connect.rb:7
```

`bottle :unneeded` has been a no-op for some time now, so all we need to do here is delete it.