Fix completion function generation
The completion functions for gds-cli are associated with the
binaries using the filename of the called binary. ie. if you
renamed 'gds' to 'umbrella', the generated function would be
associated with the binary 'umbrella', and would therefore not
trigger when `gds` was called.

We have been associating both `_gds` and `_gds-cli` with `gds`,
and so they are not correctly associated.

Now, create completion functions for `gds` by calling the binary,
and functions for `gds-cli` by calling the symlink.

Additionally, the individual `$SHELL-completion` functions have
been deprecated in alphagov/gds-cli#601, so I have updated the
syntax in the formula.